feat(GoogleOAuth): Implements google OAuth2.0 Removes commented out notification code in Navbar.js component
Updates Navbar component with User info populating avatar, implements sign in button (@navbar avatar) if user not signed in. Enables logout functionality at navbar toggle menu

Implements LOGOUT and AUTH actionTypes

Implements Auth reducers
