+++ Client/.gitignore
@@ -1,3 +1,4 @@
 node_modules/
 build/
 .DS_Store
+.env
