+++ Client/src/constants/actionTypes.js
@@ -7,3 +7,6 @@ export const CREATE = 'CREATE';
 export const UPDATE = 'UPDATE';
 export const DELETE = 'DELETE';
 export const FETCH_ALL = 'FETCH_ALL';
+
+export const AUTH = 'AUTH';
+export const LOGOUT =  'LOGOUT';
