+++ Client/src/pages/HomePage.js
@@ -4,8 +4,8 @@ import { Routes } from "../routes";
 
 //* pages
 import Trips from "./dashboard/Trips";
-// import Signin from "./examples/Signin";
-// import Signup from "./examples/Signup";
+import Signin from "./Signin";
+import Signup from "./Signup";
 import ForgotPassword from "./examples/ForgotPassword";
 import ResetPassword from "./examples/ResetPassword";
 import Lock from "./examples/Lock";
@@ -24,61 +24,61 @@ import Preloader from "../components/Preloader";
 
 
 const RouteWithLoader = ({ component: Component, ...rest }) => {
-  const [loaded, setLoaded] = useState(false);
+        const [loaded, setLoaded] = useState(false);
 
-  useEffect(() => {
-    const timer = setTimeout(() => setLoaded(true), 1000);
-    return () => clearTimeout(timer);
-  }, []);
+    useEffect(() => {
+        const timer = setTimeout(() => setLoaded(true), 1000);
+        return () => clearTimeout(timer);
+    }, []);
 
-  return (
-    <Route {...rest} render={props => ( <> <Preloader show={loaded ? false : true} /> <Component {...props} /> </> ) } />
-  );
+    return (
+        <Route {...rest} render={props => ( <> <Preloader show={loaded ? false : true} /> <Component {...props} /> </> ) } />
+    );
 };
 
 const RouteWithSidebar = ({ component: Component, ...rest }) => {
-  const [loaded, setLoaded] = useState(false);
+const [loaded, setLoaded] = useState(false);
 
-  useEffect(() => {
-    const timer = setTimeout(() => setLoaded(true), 1000);
-    return () => clearTimeout(timer);
-  }, []);
+    useEffect(() => {
+        const timer = setTimeout(() => setLoaded(true), 1000);
+        return () => clearTimeout(timer);
+    }, []);
 
 
-  return (
-    <Route {...rest} render={props => (
-      <>
-        <Preloader show={loaded ? false : true} />
-        <Sidebar />
+    return (
+        <Route {...rest} render={props => (
+        <>
+            <Preloader show={loaded ? false : true} />
+            <Sidebar />
 
-        <main className="content">
-          <Navbar />
-          <Component {...props} />
-        </main>
-      </>
-    )}
-    />
-  );
+            <main className="content">
+            <Navbar />
+            <Component {...props} />
+            </main>
+        </>
+        )}
+        />
+    );
 };
 
 export default () => (
-  <Switch>
-    {/* <RouteWithLoader exact path={Routes.Signin.path} component={Signin} />
-    <RouteWithLoader exact path={Routes.Signup.path} component={Signup} /> */}
-    <RouteWithLoader exact path={Routes.ForgotPassword.path} component={ForgotPassword} />
-    <RouteWithLoader exact path={Routes.ResetPassword.path} component={ResetPassword} />
-    <RouteWithLoader exact path={Routes.Lock.path} component={Lock} />
-    <RouteWithLoader exact path={Routes.NotFound.path} component={NotFoundPage} />
-    <RouteWithLoader exact path={Routes.ServerError.path} component={ServerError} />
+    <Switch>
+        <RouteWithLoader exact path={Routes.Signin.path} component={Signin} />
+        <RouteWithLoader exact path={Routes.Signup.path} component={Signup} />
+        <RouteWithLoader exact path={Routes.ForgotPassword.path} component={ForgotPassword} />
+        <RouteWithLoader exact path={Routes.ResetPassword.path} component={ResetPassword} />
+        <RouteWithLoader exact path={Routes.Lock.path} component={Lock} />
+        <RouteWithLoader exact path={Routes.NotFound.path} component={NotFoundPage} />
+        <RouteWithLoader exact path={Routes.ServerError.path} component={ServerError} />
 
-    {/* pages */}
-    <RouteWithSidebar exact path={Routes.Trips.path} component={Trips} />
-    <RouteWithSidebar exact path={Routes.Users.path} component={Users} />
-    <RouteWithSidebar exact path={Routes.UserTrips.path} component={UserTrips} /> 
-    {/* <RouteWithSidebar exact path={Routes.AddTrip.path} component={AddTrip} />  */}
-    {/* <RouteWithSidebar exact path={Routes.Login.path} component={Login} />  */}
+        {/* pages */}
+        <RouteWithSidebar exact path={Routes.Trips.path} component={Trips} />
+        <RouteWithSidebar exact path={Routes.Users.path} component={Users} />
+        <RouteWithSidebar exact path={Routes.UserTrips.path} component={UserTrips} /> 
+        {/* <RouteWithSidebar exact path={Routes.AddTrip.path} component={AddTrip} />  */}
+        {/* <RouteWithSidebar exact path={Routes.Login.path} component={Login} />  */}
 
 
-    <Redirect to={Routes.NotFound.path} />
-  </Switch>
+        <Redirect to={Routes.NotFound.path} />
+    </Switch>
 );
