+++ Client/src/reducers/index.js
@@ -1,5 +1,6 @@
 import { combineReducers } from "redux";
 
 import trips from './trips';
+import auth from './auth';
 
-export default combineReducers({ trips });
+export default combineReducers({ trips, auth });
