+++ Client/src/routes.js
@@ -7,8 +7,8 @@ export const Routes = {
     UserTrips: { path: "/usertrips" },
     AddTrip: { path: "/addtrip" },
     // Login: {path: "/login"},
-    // Signin: { path: "/examples/sign-in" },
-    Signup: { path: "/examples/sign-up" },
+    Signin: { path: "/sign-in" },
+    Signup: { path: "/sign-up" },
     ForgotPassword: { path: "/examples/forgot-password" },
     ResetPassword: { path: "/examples/reset-password" },
     Lock: { path: "/examples/lock" },
